pyproject.toml
[tool.poetry]
name = "narva-stat"
version = "0.1.0"
description = ""
authors = ["waroma <razer30606@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.37.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
langchain = "^0.3.3"
langchain-openai = "^0.2.2"
langchain-community = "^0.3.2"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
